Updated how student profile looks
The two cards are put next to each other. I just did this for trial on student profile page. If you guys want to go back to the old design, it is the "profile_trial" file. If you like the new design, we can change the rest of the profile pages as well.